+++ javascripts/components/index.js
@@ -1,4 +1,3 @@
-import "./celestials/navigation.js";
 import "./celestials/subtabs/teresa/teresa-tab.js";
 import "./celestials/subtabs/effarig/effarig-tab.js";
 import "./celestials/subtabs/v-tab.js";
