+++ src/components/tabs/index.js
@@ -31,6 +31,7 @@ import ImaginaryUpgradesTab from "./imaginary-upgrades/ImaginaryUpgradesTab";
 import PerksTab from "./perks/PerksTab";
 import BlackHoleTab from "./black-hole/BlackHoleTab";
 import AlchemyTab from "./alchemy/AlchemyTab";
+import CelestialNavigationTab from "./celestial-navigation/CelestialNavigationTab";
 
 const TabComponents = {
   AntimatterDimensionsTab,
@@ -64,7 +65,8 @@ const TabComponents = {
   ImaginaryUpgradesTab,
   PerksTab,
   BlackHoleTab,
-  AlchemyTab
+  AlchemyTab,
+  CelestialNavigationTab,
 };
 
 export default TabComponents;
